# gateway/openapi.yaml
openapi: 3.0.3
info:
  title: Healthcare Platform API
  description: |
    Comprehensive Healthcare Platform API for managing patients, providers, 
    appointments, claims, payments, and more.
    
    ## Authentication
    Most endpoints require JWT authentication. Include the token in the 
    `Authorization` header as `Bearer <token>`.
    
    ## Rate Limits
    - Anonymous endpoints: 30 requests/minute
    - Authenticated endpoints: 100 requests/minute
    - Admin endpoints: 30 requests/minute
    
    ## Versioning
    API version is specified in the URL path (e.g., `/api/v1/...`) or 
    via the `X-API-Version` header.
  version: 1.0.0
  contact:
    name: Healthcare Platform Support
    email: support@healthcare.com
    url: https://healthcare.com/support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.healthcare.com/v1
    description: Production server
  - url: https://staging-api.healthcare.com/v1
    description: Staging server
  - url: http://localhost:3000/api
    description: Development server
...
paths:
...
  /appointments/{id}:
    get:
      tags: [Appointments]
      summary: Get appointment details
      operationId: getAppointment
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Appointment details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Appointment'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Appointments]
      summary: Update appointment
      operationId: updateAppointment
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAppointmentRequest'
      responses:
        '200':
          description: Appointment updated
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Appointments]
      summary: Cancel appointment
      operationId: cancelAppointment
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Appointment cancelled

  # ==================== Claims ====================
...
  /payments:
    get:
      tags: [Payments]
      summary: List payments
      operationId: listPayments
      security:
        - bearerAuth: []
      parameters:
        - name: patient_id
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, completed, failed, refunded]
      responses:
        '200':
          description: List of payments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentListResponse'

    post:
      tags: [Payments]
      summary: Process a payment
      operationId: processPayment
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProcessPaymentRequest'
      responses:
        '201':
          description: Payment processed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'

  # ==================== Reports ====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login

    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for external integrations
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
        code:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            type: object
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        pages:
          type: integer
...
    Appointment:
      type: object
      properties:
        id:
          type: integer
        patient_id:
          type: integer
        provider_id:
          type: integer
        provider_name:
          type: string
        provider_specialty:
          type: string
        appointment_date:
          type: string
          format: date-time
        duration_minutes:
          type: integer
        appointment_type:
          type: string
        status:
          type: string
          enum: [scheduled, confirmed, completed, cancelled, no_show]
        notes:
          type: string
        reason_for_visit:
          type: string
        virtual:
          type: boolean
        meeting_link:
          type: string
        created_at:
          type: string
          format: date-time
...
    UpdateAppointmentRequest:
      type: object
      properties:
        appointment_date:
          type: string
          format: date-time
        duration_minutes:
          type: integer
        status:
          type: string
          enum: [scheduled, confirmed, completed, cancelled, no_show]
        notes:
          type: string
...
    Payment:
      type: object
      properties:
        id:
          type: integer
        patient_id:
          type: integer
        amount:
          type: number
        currency:
          type: string
          default: USD
        status:
          type: string
          enum: [pending, completed, failed, refunded]
        payment_method:
          type: string
        transaction_id:
          type: string
        created_at:
          type: string
          format: date-time
    PaymentListResponse:
      type: object
      properties:
        payments:
          type: array
          items:
            $ref: '#/components/schemas/Payment'
        pagination:
          $ref: '#/components/schemas/Pagination'
    ProcessPaymentRequest:
      type: object
      required: [patient_id, amount, payment_method]
      properties:
        patient_id:
          type: integer
        amount:
          type: number
        currency:
          type: string
          default: USD
        payment_method:
          type: string
        appointment_id:
          type: integer
        metadata:
          type: object
...